Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 1002ms, Memory: 0MB, Status: Time Limit Exceed - CodeTree

+++ 240129/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.cpp
@@ -8,7 +8,7 @@ long long a;
 long long b, c;
 vector<long long> v;
 
-long long func(int x){
+long long func(long long x){
     long long left = 1;
     long long right = a;
     long long cnt = 0;
